+++ examples/cite-service/src/main.rs
@@ -1,33 +1,32 @@
-use clap::Parser;
-use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
-
 use ogcapi::{
     processes::echo::Echo,
-    services::{AppState, Config, Service},
+    services::{AppState, Config, ConfigParser, Drivers, Service},
 };
 
 #[tokio::main]
 async fn main() {
     // setup env
-    dotenvy::dotenv().ok();
+    ogcapi::services::setup_env();
 
     // setup tracing
-    tracing_subscriber::registry()
-        .with(tracing_subscriber::EnvFilter::new(
-            "cite_service=debug,ogcapi=debug,sqlx=warn",
-        ))
-        .with(tracing_subscriber::fmt::layer().pretty())
-        .init();
+    ogcapi::services::telemetry::init();
 
     // Config
-    let config = Config::parse();
+    let config = Config::try_parse().unwrap();
+
+    // Drivers
+    let drivers = Drivers::try_new_from_env().await.unwrap();
 
     // Application state
-    let state = AppState::new_from(&config).await;
+    let state = AppState::new(drivers).await;
 
     // Register processes/processors
     let state = state.processors(vec![Box::new(Echo)]);
 
     // Build & run with hyper
-    Service::new_with(&config, state).await.serve().await;
+    Service::try_new_with(&config, state)
+        .await
+        .unwrap()
+        .serve()
+        .await;
 }