+++ ogcapi-services/src/lib.rs
@@ -13,9 +13,13 @@ pub use config::Config;
 pub use error::Error;
 pub use openapi::ApiDoc;
 pub use service::Service;
-pub use state::AppState;
+pub use state::{AppState, Drivers};
 
 #[doc(hidden)]
 pub use clap::Parser as ConfigParser;
 
 pub type Result<T, E = Error> = std::result::Result<T, E>;
+
+pub fn setup_env() {
+    dotenvy::dotenv().ok();
+}